Update readme
This application is built with rails and tested with rspec. I initially wanted to try another api like grape or use rails-api but couldn't configure rspec to work properly. To run the specs type rspec spec/ once you're inside the root of the application. Distances are calculated using the haversine formula.